src/main/java/Client.java
import org.sql2o.*;

public class Client {
  private String name;
  private String description;
  private int id;

  public Client(String name, String description) {
    this.name = name;
    this.description = description;
  }

  public String getName() {
    return name;
  }

  public String getDescription() {
    return description;
  }

  public int getId() {
    return id;
  }

  @Override
  public boolean equals(Object otherClient) {
    if (!(otherClient instanceof Client)) {
      return false;
    } else {
      Client newClient = (Client) otherClient;
      return this.getName().equals(newClient.getName()) && this.getDescription().equals(newClient.getDescription());
    }
  }

  public void save(){
    String sql = "INSERT INTO clients (name, description) VALUES (:name, :description)";
    try(Connection con = DB.sql2o.open()){
      this.id = (int) con.createQuery(sql, true)
        .addParameter("name", this.name)
        .addParameter("description", this.description)
        .executeUpdate()
        .getKey();
    }
  }


}
